Receiver Token
Receiver token now creates a token.txt the first time it receives data. If it the file already exists it will compare the data in the file to token value sent by sender, if token is in database it will respond with "TOKEN IS IN DATABASE", if it is not there it is meant to add it as a new connection into token.txt, alongside the IP address and port number.

I have tested out the system with another unique token value. It just replaces the first line of the token file. which is something i plan on fixing tomorrow.

I couldn't get the receivers Ip address which is why it is not stored in the same token.txt, which would have then messed up the formatting of the token.txt file, but a separate token file called mytoken.txt. It may not need to be fixed but idk right now.

+++ FileTransfer/Token/receiver_token.py
@@ -1,29 +1,57 @@
 from socket import *
 import sys
-
+import os.path
 
 serverPort =12075
 SEPARATOR = "<SEPARATOR>"
 mytoken=10001
-serverSocket = socket (AF_INET,SOCK_STREAM)
 
+
+
+
+serverSocket = socket (AF_INET,SOCK_STREAM)
 serverSocket.bind (('',serverPort))
+
+with open ("mytoken.txt","w") as my:
+        print(mytoken,file=my)
+        my.close()
+
 serverSocket.listen(1)
 print("the server is ready to recieve")
 
 #ipAddr=serverSocket.getsockname()
 
 #def mydata()
-
-
+#i dont know what your path is
+path =("C:\\CSU23021\\Comp_Networks_Project_2\\FileTransfer\\Token\\token.txt")
+#print(os.path.isfile(path))
 while 1:
     connectionSocket,addr = serverSocket.accept()
     received = connectionSocket.recv(65536).decode()
     port, token = received.split(SEPARATOR)
-    print (addr[0], port, token)
-    with open("token.txt","w") as f:
-       sys.stdout = f
-       print("Ip",addr[0],"Port", port, "Token",token)
-    exit()
-
+    #if token doesnt exist write here rn
+    if os.path.isfile(path) != True:
+        with open("token.txt","w") as f:
+            print(addr[0],",", port,",",token,file=f)
+            f.close
+            exit()
+    #else if it does exist read token values and compare against ones in db,if it doesnt exist add it.
+    else:
+        with open ('token.txt',"r+") as f:
+            content =f.readline()
+            #print (content)
+            contentsplit = content.split(",")
+            tkn = int(contentsplit[2])
+            print(tkn)
+            print(int(token))
+            if int(token) == tkn :
+                print("Token is in database")
+                f.close  
+        
+    with open("token.txt","r+") as f:
+        print(addr[0],",", port,",",token,file=f)
+        f.close
+        print("token.txt was written to")
+exit()
+    
 serverSocket.close
